requirements.txt
Flask
Flasgger
python-dotenv
mysql-connector-python
requests
PyJWT